+++ services/web-app/pages/assets/[host]/[org]/[repo]/[library]/[ref]/[asset].js
@@ -201,9 +201,9 @@ const Asset = ({ libraryData, params }) => {
                       <dd className={dashboardStyles.meta}>
                         <StatusIcon
                           className={styles['status-icon']}
-                          status={assetData.content.status}
+                          status={assetData.statusKey}
                         />
-                        {get(status, `[${assetData.content.status}].name`, '–')}
+                        {status[assetData.statusKey]?.name || '-'}
                       </dd>
                     </Column>
                     <Column